Move the Skill Pill component to the Home Componenets directory
[RM-29]
